[DS-2377] Move tests to match tested classes
