+++ dspace-services/src/test/java/org/dspace/kernel/DSpaceKernelManagerTest.java
@@ -5,13 +5,10 @@
  *
  * http://www.dspace.org/license/
  */
-package org.dspace.servicemanager;
+package org.dspace.kernel;
 
-import org.dspace.kernel.DSpaceKernelImpl;
 import static org.junit.Assert.*;
 
-import org.dspace.kernel.DSpaceKernel;
-import org.dspace.kernel.DSpaceKernelManager;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
